feat: Implement push monitor heartbeat and incident management
- Added dedicated `last_heartbeat_at` column to `monitors` for push monitors to track heartbeat separately from scheduled checks.
- Updated `RecordPushHeartbeat` SQL function to handle push monitor heartbeats and status updates.
- Modified incident lifecycle to handle push monitor status changes, ensuring incidents are created for overdue checks and resolved correctly on recovery.
- Created tests for push incident lifecycle and scheduled overdue scenarios to validate new functionality.
- Updated API documentation to reflect new push monitor endpoints and heartbeat semantics.
- Refactored related services and tests to accommodate changes in the incident and audit logging systems.

+++ backend/internal/db/status_pages.sql.go
@@ -227,7 +227,7 @@ const monitorUptimeWindow = `SELECT
     COUNT(*) AS total_count
 FROM monitor_results
 WHERE monitor_id = $1
-  AND checked_at >= NOW() - ($2 || ' seconds')::interval`
+  AND checked_at >= NOW() - ($2::int || ' seconds')::interval`
 
 type UptimeCounts struct {
 	UpCount    int64
